+++ cypress/support/commands.ts
@@ -4,13 +4,12 @@
 // existing commands.
 //
 // For more comprehensive examples of custom
-// commands please read more here:
+//  commands, please read more here:
 // https://on.cypress.io/custom-commands
 // ***********************************************
 //
 //
 import {isEqual} from 'lodash';
-import {searchFHIRServer} from "./mocks/fhir-server-mocks";
 import {CypressUtil} from "./cypress-util";
 // -- This is a parent command --
 // Cypress.Commands.add('login', (email, password) => { ... })
@@ -28,7 +27,7 @@ import {CypressUtil} from "./cypress-util";
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
 
 /**
- * Load home page and wait until LForms is loaded.
+ * Load the home page and wait until LForms is loaded.
  */
 Cypress.Commands.add('loadHomePage',() => {
   cy.clearSession();
@@ -47,7 +46,7 @@ Cypress.Commands.add('loadHomePageWithLoincOnly',() => {
 
 
 /**
- * Visit home page and assert LForms, but do not deal with LOINC notice.
+ * Visit the home page and assert LForms, but do not deal with LOINC notice.
  */
 Cypress.Commands.add('goToHomePage', () => {
   CypressUtil.mockLFormsLoader();
@@ -86,11 +85,12 @@ Cypress.Commands.add('clearSession',() => {
 
 
 /**
- * Get an item from local storage.
+ * Read a local storage item.
  */
-Cypress.Commands.add('getLocalStorageItem',(item) => {
+Cypress.Commands.add('getLocalStorageItem', (itemName) => {
   return cy.window()
-    .its('localStorage').invoke('getItem', item);
+    .its('localStorage')
+    .invoke('getItem', itemName);
 });
 
 
@@ -109,7 +109,7 @@ Cypress.Commands.add('getSessionStorageItem',(item) => {
  * @param fileName - Name of the file to upload
  */
 Cypress.Commands.add('uploadFile',(fileName, handleWarning) => {
-  cy.fixture(fileName, { encoding: null }).as('myFixture');
+  cy.fixture(fileName, null).as("myFixture");
   cy.get('input[type="file"]').selectFile('@myFixture', {force: true});
   if(handleWarning) {
     cy.handleWarning();
@@ -117,7 +117,7 @@ Cypress.Commands.add('uploadFile',(fileName, handleWarning) => {
 });
 
 /**
- * Command to get json from 'Preview'
+ * Command to get JSON from 'Preview'
  */
 Cypress.Commands.add('questionnaireJSON', () => {
   return CypressUtil.getQuestionnaireJSON();
@@ -127,15 +127,15 @@ Cypress.Commands.add('questionnaireJSON', () => {
  * Command to select data type in item editor.
  */
 Cypress.Commands.add('selectDataType', (type) => {
-  cy.get('#type').select(type);
+  cy.getItemTypeField().select(type);
 });
 
 /**
- * Select a node by its text in the sidebar. The text is read from tooltip.
+ * Select a node by its text in the sidebar. The text is read from the tooltip.
  */
 Cypress.Commands.add('getTreeNode', (text) => {
   return cy.get('div[role="tooltip"]:contains("'+text+'")').invoke('attr', 'id').then((tooltipId) => {
-    return cy.get('div[aria-describedby="' + tooltipId + '"]').should('be.visible');
+    return cy.get('div[aria-describedby="' + tooltipId + '"]:visible');
   });
 });
 
@@ -143,22 +143,23 @@ Cypress.Commands.add('getTreeNode', (text) => {
  * Toggle expansion and collapse of tree node having children.
  */
 Cypress.Commands.add('toggleTreeNodeExpansion', (text) => {
-  const tooltipId = cy.get('div[role="tooltip"]:contains("'+text+'")').invoke('attr', 'id').then((tooltipId) => {
+  cy.get('div[role="tooltip"]:contains("'+text+'")').invoke('attr', 'id').then(tooltipId => {
     cy.get('tree-root tree-viewport tree-node-collection tree-node tree-node-wrapper div.node-wrapper div tree-node-content div')
       .filter('div[aria-describedby="'+tooltipId+'"]').parents('div.node-wrapper').find('tree-node-expander').as('expander');
     cy.get('@expander').should('be.visible');
     cy.get('@expander').click();
     cy.getTreeNode(text).should('be.visible');
-  }, (err)=> {console.error(err)});
+  });
 });
 
 
 /**
  * Load LOINC form using a search term. Picks first item from the result list.
- * @param searchTerm - Search term to search LOINC database.
+ * @param searchTerm - Search term to search the LOINC database.
  */
 Cypress.Commands.add('loadLOINCForm', (searchTerm) => {
-  cy.contains('nav.navbar button', 'Import').scrollIntoView().click();
+  cy.contains('nav.navbar button', 'Import').as('importBtn').scrollIntoView();
+  cy.get('@importBtn').click();
   cy.get('div.dropdown-menu.show form input[placeholder="Search LOINC"]').as('searchBox');
   cy.get('@searchBox').type(searchTerm);
   cy.get('ngb-typeahead-window').should('be.visible');
@@ -168,7 +169,7 @@ Cypress.Commands.add('loadLOINCForm', (searchTerm) => {
 });
 
 /**
- * Get json from FHIR server response after create/update interaction.
+ * Get JSON from FHIR server response after create/update interaction.
  * @param menuText - Menu text to pick the menu item.
  */
 Cypress.Commands.add('FHIRServerResponse', (menuText, serverBaseUrl = 'https://lforms-fhir.nlm.nih.gov/baseR4') => {
@@ -235,7 +236,7 @@ Cypress.Commands.add('addAnswerOptions', () => {
   cy.get('[id^="answerOption.1.valueCoding.code"]').type('c2');
   cy.get('[id^="answerOption.1.valueCoding.system"]').type('s2');
   cy.get('[id^="answerOption.1.valueCoding.__$score"]').type('3');
-  // Select first option
+  // Select the first option
   cy.contains('div', 'Value method').find('[for^="__$valueMethod_pick-initial"]').click();
   cy.get('[id^="pick-answer"]').as('pickAnswer');
   cy.get('@pickAnswer').click();
@@ -270,7 +271,7 @@ Cypress.Commands.add('addAnswerOptions', () => {
  * Test code yes no options
  */
 Cypress.Commands.add('includeExcludeCodeField', {prevSubject: true}, (codeOptionElement, formOrItem) => {
-  const formTesting = formOrItem === 'form' ? true : false;
+  const formTesting = formOrItem === 'form';
   cy.wrap(codeOptionElement).find('[for^="booleanRadio_true"]').as('codeYes');
   cy.wrap(codeOptionElement).find('[for^="booleanRadio_false"]').as('codeNo');
   cy.get('[id^="booleanRadio_false"]').should('be.checked');
@@ -283,9 +284,9 @@ Cypress.Commands.add('includeExcludeCodeField', {prevSubject: true}, (codeOption
   cy.get('@codeYes').click();
   cy.get('[id^="code.0.code_"]').as('code');
   cy.get('@code').type('ab ');
-  cy.get('@code').next('small')
-    .should('be.visible')
-    .contains('Spaces are not allowed at the beginning or end.');
+  cy.get('@code').next('ul').find('small').as('codeError');
+  cy.get('@codeError').should('be.visible');
+  cy.get('@codeError').contains('Spaces are not allowed at the beginning or end.');
   cy.get('@code').clear();
   cy.get('@code').type(coding.code);
   cy.get('[id^="code.0.system_"]').type(coding.system);
@@ -318,9 +319,9 @@ Cypress.Commands.add('dragAndDropNode', (dragNodeText, dropNodeText) => {
 
   const dropSelector = '.node-content-wrapper span:contains("' + dropNodeText + '")';
   const dragSelector = '.node-content-wrapper span:contains("' + dragNodeText + '")';
-  let droppable, coords;
+  let coords: any;
   cy.get(dropSelector).should(($eList) => {
-    droppable = $eList[0];
+    const droppable = $eList[0];
     coords = droppable.getBoundingClientRect();
   });
 
@@ -345,12 +346,15 @@ Cypress.Commands.add('dragAndDropNode', (dragNodeText, dropNodeText) => {
 
 /**
  * Interact with FHIR server selection and do search with <code>titleSearchTerm</code>
- * and pick first result to load into the form builder.
- * Make sure to create mock response based on titleSearchTerm.
+ * and pick the first result to load into the form builder.
+ * Make sure to create a mock response based on titleSearchTerm.
  */
 Cypress.Commands.add('fhirSearch', (titleSearchTerm) => {
-  searchFHIRServer(titleSearchTerm,
-    `fhir-server-mock-response-${titleSearchTerm}.json`);
+
+  cy.intercept(
+    `**title:contains=${titleSearchTerm}**`,
+    {fixture: `fhir-server-mock-response-${titleSearchTerm}.json`}).as("searchFHIRServer");
+
   cy.get('input[type="radio"][name="fhirServer"]').first().click();
   cy.contains('div.modal-footer button', 'Continue').click();
   cy.get('input.form-control[placeholder="Search any text field"]').type(titleSearchTerm);
@@ -362,7 +366,7 @@ Cypress.Commands.add('fhirSearch', (titleSearchTerm) => {
 });
 
 /**
- * Expect warning dialog and click continue.
+ * Expect the warning dialog and click continue.
  */
 Cypress.Commands.add('handleWarning', () => {
   cy.contains('.modal-title', 'Replace existing form?').should('be.visible');
@@ -370,7 +374,7 @@ Cypress.Commands.add('handleWarning', () => {
 });
 
 /**
- * Read form from local storage and compare it with default form.
+ * Read a form from local storage and compare it with the default form.
  * Yields boolean
  */
 Cypress.Commands.add('isDefault', () => {
@@ -394,23 +398,12 @@ Cypress.Commands.add('getCurrentForm', () => {
   return cy.getLocalStorageItem('fhirQuestionnaire').then((formStr) => {
     const form = formStr && formStr.length > 0 ? JSON.parse(formStr) : null;
     return cy.wrap(form);
-  }, (err) => {
-    return err;
   });
 });
 
 /**
- * Read a local storage item.
- */
-Cypress.Commands.add('getLocalStorageItem', (itemName) => {
-  return cy.window()
-    .its('localStorage')
-    .invoke('getItem', itemName);
-});
-
-/**
  * Reset form builder.
- * Using Close menu option to reset.
+ * Using the Close menu option to reset.
  */
 Cypress.Commands.add('resetForm', () => {
   cy.contains('nav.navbar > div > button', 'Close').click();
@@ -464,7 +457,7 @@ Cypress.Commands.add('editableLinkId', () => {
  * @param errorMessage - the error message to validate.
  */
 Cypress.Commands.add('checkLinkIdErrorIsDisplayed', (errorMessage) => {
-  // The link id text input should be outline in red
+  // The link id text input should be outlined in red
   cy.editableLinkId()
     .should('have.class', 'invalid');
      // The error message should display at the bottom of the text input
@@ -483,7 +476,7 @@ Cypress.Commands.add('checkLinkIdErrorIsDisplayed', (errorMessage) => {
  * Check whether the error message for the linkId field is no longer displayed in the UI.
  */
 Cypress.Commands.add('checkLinkIdErrorIsNotDisplayed', () => {
-  // The link id text input should not have outline in red
+  // The link id text input should not have an outline in red
   cy.editableLinkId()
     .should('not.have.class', 'invalid');
   // The error message should not be displayed at the bottom of the text input
@@ -520,8 +513,8 @@ Cypress.Commands.add('booleanFieldClick', (fieldLabel, rbValue) => {
 /**
  * Get a radio button identified by the label of the group and the label of the radio button.
  *
- * Use radio input element to make assertions on input status, such as selected or not. It is not
- * suitable for mouse actions as it is hidden from mouse-pointer. Instead use its label to perform
+ * Use a radio input element to make assertions on input status, such as selected or not. It is not
+ * suitable for mouse actions as it is hidden from a mouse-pointer, instead use its label to perform
  * mouse actions.
 */
 
@@ -557,7 +550,7 @@ Cypress.Commands.add('getBooleanFieldParent', (fieldLabel) => {
 
 /**
  * Get parent for elements of boolean input/label in initial[x].valueBoolean field.
- * The initial value field has different css path compared to above general boolean field.
+ * The initial value field has a different CSS path compared to the above general boolean field.
  */
 Cypress.Commands.add('getInitialValueBooleanParent', () => {
   return cy.get('lfb-table lfb-label label').contains('Initial value').parent().parent().next()
@@ -565,43 +558,79 @@ Cypress.Commands.add('getInitialValueBooleanParent', () => {
 });
 
 /**
- * Get input element for 'Initial value' boolean field.
+ * Get an input element for 'Initial value' boolean field.
  */
 Cypress.Commands.add('getInitialValueBooleanInput', (rbValue) => {
   return cy.getInitialValueBooleanParent().find('input[id^="booleanRadio_'+rbValue+'"]');
 });
 
 /**
- * Click radio button of 'Initial value' boolean field.
+ * Click a radio button of 'Initial value' boolean field.
  */
 Cypress.Commands.add('getInitialValueBooleanClick', (rbValue) => {
-  return getInitialValueBooleanParent().find('label[for^="booleanRadio_'+rbValue+'"]').click();
+  return cy.getInitialValueBooleanParent().find('label[for^="booleanRadio_'+rbValue+'"]').click();
 });
 
 /**
- * Click radio button for the 'Type initial value' boolean field under the Value Method.
+ * Get the input element using its label text.
+ * Works provided the label[for] === input[id].
+ * @param parentSelector - The parent selector to search within. Helps to constrain
+ * the search to a specific part of the form.
+ * @param label - The label text to find the input element.
  */
-Cypress.Commands.add('getTypeInitialValueValueMethodClick', (rbValue) => {
+Cypress.Commands.add('getByLabel', (parentSelector: string, label: string) => {
+  return cy.get(parentSelector)
+    .find('label')
+    .contains(label).invoke('attr', 'for').then((id) => {
+      return cy.get('#' + id);
+    });
+});
+
+/**
+ * Get the title field from the form level page.
+ */
+Cypress.Commands.add('getFormTitleField', () => {
+  return cy.getByLabel('lfb-form-fields', 'Title');
+});
+
+/**
+ * Get the data type field from the item editor.
+ */
+Cypress.Commands.add('getItemTypeField', () => {
+  return cy.getByLabel('lfb-ngx-schema-form', 'Data type');
+});
+
+/**
+ * Get the question text field from the item editor.
+ */
+Cypress.Commands.add('getItemTextField', () => {
+  return cy.getByLabel('lfb-ngx-schema-form', 'Question text');
+});
+
+/**
+ * Click a radio button for the 'Type initial value' boolean field under the Value Method.
+ */
+Cypress.Commands.add('getTypeInitialValueValueMethodClick', () => {
   return cy.contains('div', 'Value method').find('[for^="__$valueMethod_type-initial"]').click();
 });
 
 /**
- * Click radio button for the 'Pick initial value' boolean field under the Value Method.
+ * Click a radio button for the 'Pick initial value' boolean field under the Value Method.
  */
-Cypress.Commands.add('getPickInitialValueValueMethodClick', (rbValue) => {
+Cypress.Commands.add('getPickInitialValueValueMethodClick', () => {
   return cy.contains('div', 'Value method').find('[for^="__$valueMethod_pick-initial"]').click();
 });
 /**
- * Click radio button for the 'Compute initial value' boolean field under the Value Method.
+ * Click a radio button for the 'Compute initial value' boolean field under the Value Method.
  */
-Cypress.Commands.add('getComputeInitialValueValueMethodClick', (rbValue) => {
+Cypress.Commands.add('getComputeInitialValueValueMethodClick', () => {
   return cy.contains('div', 'Value method').find('[for^="__$valueMethod_compute-initial"]').click();
 });
 
 /**
- * Click radio button for the 'Continuously compute value' boolean field under the Value Method.
+ * Click a radio button for the 'Continuously compute value' boolean field under the Value Method.
  */
-Cypress.Commands.add('getComputeContinuouslyValueValueMethodClick', (rbValue) => {
+Cypress.Commands.add('getComputeContinuouslyValueValueMethodClick', () => {
   return cy.contains('div', 'Value method').find('[for^="__$valueMethod_compute-continuously"]').click();
 });
 
@@ -619,8 +648,8 @@ Cypress.Commands.add('getComputeContinuouslyValueValueMethodClick', (rbValue) =>
  * optionally checks the number of results, selects an option using special key sequences,
  * and then runs the provided assertion logic.
  */
-function handleAutocomplete(autocompleteElement, clearBeforeTyping, searchKeyword, expectedListSize,
-                            specialCharacterSequencesText, assertionFn: () => void) {
+function handleAutocomplete(autocompleteElement: JQuery<HTMLInputElement>, clearBeforeTyping: boolean, searchKeyword: string, expectedListSize: number,
+                            specialCharacterSequencesText: string, assertionFn: () => void) {
   if (clearBeforeTyping) {
     cy.wrap(autocompleteElement).clear();
   }
@@ -660,7 +689,7 @@ function handleAutocomplete(autocompleteElement, clearBeforeTyping, searchKeywor
  * @param searchKeyword - (Optional) The keyword to type into the autocomplete input.
  * @param expectedListSize - (Optional) The expected number of options in the search results. Pass null to skip this check.
  * @param specialCharacterSequencesText - Special key sequences to send (e.g., '{downarrow}{enter}').
- * @param expectedResults - (Optional) The expected array of text for the selected results to assert.
+ * @param expectedResults - (Optional) The expected array of texts for the selected results to assert.
  *
  * This command types an optional keyword into a multi-select autocomplete input, waits for the search results to appear,
  * optionally checks the number of results, selects an option using special key sequences,
@@ -709,7 +738,7 @@ Cypress.Commands.add('selectAutocompleteOptions',
  *
  * This command types an optional keyword into an autocomplete input, waits for the search results to appear,
  * optionally checks the number of results, selects an option using special key sequences,
- * and verifies that the expected result appear in the selection display.
+ * and verifies that the expected result appears in the selection display.
  */
 Cypress.Commands.add('selectAutocompleteOption',
   (autocompleteElement, clearBeforeTyping, searchKeyword, expectedListSize, specialCharacterSequencesText,
@@ -740,7 +769,6 @@ Cypress.Commands.add('selectAutocompleteOption',
   }
 );
 
-
 /**
  * Custom command to check the Question Item Control UI logic for a given data type and expected UI state.
  * @param {string} type - The data type to select (e.g., 'integer').
@@ -762,7 +790,7 @@ Cypress.Commands.add('checkQuestionItemControlUI',
   const answerValueSetOptionsVisible = itemControlOptionsAfterAnswerValueSet;
 
   // Select the data type
-  cy.get('#type').select(type);
+  cy.selectDataType(type);
 
   // Check 'Create answer list' label visibility
   if (createAnswerListVisible) {
@@ -825,3 +853,71 @@ Cypress.Commands.add('checkQuestionItemControlUI',
     }
   }
 });
+// Helps remove TypeScript errors and auto completing the Cypress commands in TypeScript
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      loadHomePage(): Chainable<void>;
+      loadHomePageWithLoincOnly(): Chainable<void>;
+      goToHomePage(): Chainable<void>;
+      acceptAllTermsOfUse(): Chainable<void>;
+      acceptLoincOnly(): Chainable<void>;
+      clearSession(): Chainable<void>;
+      uploadFile(fileName: string, handleWarning?: boolean): Chainable<void>;
+      questionnaireJSON(): Chainable<any>;
+      selectDataType(type: string): Chainable<void>;
+      toggleTreeNodeExpansion(text: string): Chainable<JQuery<HTMLElement>>;
+      loadLOINCForm(searchTerm: string): Chainable<void>;
+      FHIRServerResponse(menuText: string, serverBaseUrl?: string): Chainable<any>;
+      enterAnswerOptions(codings: any[]): Chainable<void>;
+      addAnswerOptions(): Chainable<void>;
+      includeExcludeCodeField(formOrItem: 'form' | 'item'): Chainable<void>;
+      dragAndDropNode(dragNodeText: string, dropNodeText: string): Chainable<JQuery<HTMLElement>>;
+      fhirSearch(titleSearchTerm: string): Chainable<JQuery<HTMLElement>>;
+      handleWarning(): Chainable<JQuery<HTMLElement>>;
+      isDefault(): Chainable<boolean>;
+      resetForm(): Chainable<void>;
+      waitForSpinner(): Chainable<JQuery<HTMLElement>>;
+      clickTreeNode(nodeText: string): Chainable<JQuery<HTMLElement>>;
+      getExtensions(extensionsArray: any[], url: string): any[];
+      tsUrl(): Cypress.Chainable<JQuery<HTMLElement>>;
+      editableLinkId(): Cypress.Chainable<JQuery<HTMLElement>>;
+      checkLinkIdErrorIsDisplayed(errorMessage: string): Cypress.Chainable<JQuery<HTMLElement>>;
+      checkLinkIdErrorIsNotDisplayed(): Cypress.Chainable<JQuery<HTMLElement>>;
+      expandAdvancedFields(): Cypress.Chainable<JQuery<HTMLElement>>;
+      collapseAdvancedFields(): Cypress.Chainable<JQuery<HTMLElement>>;
+      booleanFieldClick(fieldLabel: string, rbValue: boolean): Cypress.Chainable<JQuery<HTMLElement>>;
+      getBooleanFieldParent(fieldLabel: string): Chainable<JQuery<HTMLElement>>;
+      getBooleanInput(fieldLabel: string, rbValue: boolean): Chainable<JQuery<HTMLElement>>;
+      getByLabel(parentSelector: string, label: string): Cypress.Chainable<JQuery<HTMLElement>>;
+      getCurrentForm(): Chainable<any>;
+      getFormTitleField(): Cypress.Chainable<JQuery<HTMLElement>>;
+      getInitialValueBooleanClick(rbValue: boolean): Chainable<JQuery<HTMLElement>>;
+      getInitialValueBooleanInput(rbValue: boolean): Chainable<JQuery<HTMLElement>>;
+      getInitialValueBooleanParent(): Chainable<JQuery<HTMLElement>>;
+      getItemTextField(): Cypress.Chainable<JQuery<HTMLElement>>;
+      getItemTypeField(): Cypress.Chainable<JQuery<HTMLElement>>;
+      getLocalStorageItem(itemName: string): Chainable<string | null>;
+      getRadioButton(groupLabel: string, rLabel: string): Chainable<JQuery<HTMLElement>>;
+      getRadioButtonLabel(groupLabel: string, rLabel: string): Cypress.Chainable<JQuery<HTMLElement>>;
+      getSessionStorageItem(item: string): Chainable<string | null>;
+      getTreeNode(text: string): Chainable<JQuery<HTMLElement>>;
+      getTypeInitialValueValueMethodClick(): Chainable<JQuery<HTMLElement>>;
+      getPickInitialValueValueMethodClick(): Chainable<JQuery<HTMLElement>>;
+      getComputeInitialValueValueMethodClick(): Chainable<JQuery<HTMLElement>>;
+      getComputeContinuouslyValueValueMethodClick(): Chainable<JQuery<HTMLElement>>;
+      selectAutocompleteOption(
+        autoCompleteInput: JQuery<HTMLInputElement>, clearBeforeTyping: boolean,
+        searchKeyword: string, expectedListSize: number, specialCharacterSequencesText: string,
+        expectedResult: string): Chainable<void>;
+      selectAutocompleteOptions(
+        autoCompleteInput: JQuery<HTMLInputElement>, clearBeforeTyping: boolean,
+        searchKeyword: string, expectedListSize: number, specialCharacterSequencesText: string,
+        expectedResults: string[]): Chainable<void>;
+
+      checkQuestionItemControlUI(
+        type: string, questionItemControlOptions: string[], itemControlOptions: string[],
+        itemControlOptionsAfterRepeat: string[], itemControlOptionsAfterAnswerValueSet: string[]): Chainable<void>;
+    }
+  }
+}