+++ src/app/lib/lforms-widget-registry.ts
@@ -41,6 +41,9 @@ import {TextAreaComponent} from './widgets/textarea/textarea.component';
 import {DatetimeComponent} from './widgets/datetime/datetime.component';
 import {EditableLinkIdComponent} from './widgets/editable-link-id/editable-link-id.component';
 import {HelpTextComponent} from "./widgets/help-text/help-text.component";
+import { ContainedComponent } from './widgets/contained/contained.component';
+import { ValueSetResourceComponent } from './widgets/value-set-resource/value-set-resource.component';
+
 import { ValueMethodComponent } from './widgets/value-method/value-method.component';
 import { PickAnswerComponent } from './widgets/pick-answer/pick-answer.component';
 import { ExpressionEditorComponent } from './widgets/expression-editor/expression-editor.component';
@@ -93,6 +96,8 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('quantity-unit', QuantityUnitComponent);
     this.register('terminology-server', TerminologyServerComponent);
     this.register('editable-link-id', EditableLinkIdComponent);
+    this.register("contained", ContainedComponent);
+    this.register("value-set", ValueSetResourceComponent);
     this.register('value-method', ValueMethodComponent);
     this.register('pick-answer', PickAnswerComponent);
     this.register('expression-editor', ExpressionEditorComponent);