+++ src/app/lib/widgets/expression-editor/expression-editor.component.ts
@@ -19,10 +19,10 @@ export class ExpressionEditorComponent extends LfbControlWidgetComponent impleme
   @Input() model: any;
   @Input() exp: string;
   @Output() questionnaireChange = new EventEmitter<fhir4.Questionnaire>();
-  subscriptions: Subscription[] = [];
 
   elementId: string;
-  control = new FormControl();
+  // control is already declared in the base class.
+  // control = new FormControl();
   questionnaire = null;
   private LANGUAGE_FHIRPATH = 'text/fhirpath';
   linkId: string;
@@ -30,7 +30,7 @@ export class ExpressionEditorComponent extends LfbControlWidgetComponent impleme
   valueMethod: string;
   itemId: number;
   faAdd = faPlusCircle;
-  name: string;
+  // name: string;
   /**
    * Invoke super constructor.
    *
@@ -73,7 +73,7 @@ export class ExpressionEditorComponent extends LfbControlWidgetComponent impleme
         if (!this.extensionsService.isEmptyValueExpression(outputExpression)) {
           this.expression = outputExpression?.valueExpression?.expression;
           const outputExpressionExtension = { ...outputExpression};
-          outputExpressionExtension.url = this.getUrlByValueMethod(this.valueMethod); 
+          outputExpressionExtension.url = this.getUrlByValueMethod(this.valueMethod);
 
           this.extensionsService.insertExtensionAfterURL(ExtensionsService.VARIABLE, [outputExpressionExtension]);
         }
@@ -82,7 +82,7 @@ export class ExpressionEditorComponent extends LfbControlWidgetComponent impleme
   }
 
   /**
-   * Fetches the 'initial expression' from the FormProperty. If it is empty, returns the 'calculated expression' instead. 
+   * Fetches the 'initial expression' from the FormProperty. If it is empty, returns the 'calculated expression' instead.
    * @returns - the output expression from either the 'initial expression' or the 'calculated expression'.
    */
   getOutputExpressionFromFormProperty(): any {